points-worker uses shared import loader helper
